+++ workspaces/ai-integrations/plugins/ai-experience/src/utils/constants.ts
@@ -18,108 +18,18 @@ export const LEARNING_SECTION_ITEMS = [
     title: 'Get started',
     description: 'Learn about Red Hat Developer Hub.',
     buttonText: 'Go to Tech Docs',
-    buttonLink: '#',
+    buttonLink: '/docs',
   },
   {
     title: 'Explore',
     description: 'Explore AI models, servers and templates.',
     buttonText: 'Go to Catalog',
-    buttonLink: '#',
+    buttonLink: '/catalog',
   },
   {
     title: 'Learn',
     description: 'Explore and develop new skills in AI.',
     buttonText: 'Go to Learning Paths',
-    buttonLink: '#',
-  },
-];
-
-export const AI_MODELS = [
-  {
-    title: 'granite-7b-starter',
-    version: '1.4.0',
-    description:
-      'Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model. Custom Red Hat phase00 tuned base model.',
-    tags: [
-      'LAB starter',
-      'Task_name 1',
-      'Task_name 2',
-      'Task_name 3',
-      'Task_name 4',
-    ],
-  },
-  {
-    title:
-      'prometheus-8x7b-v2-0-prometheus-8x7b-v2-0-prometheus-8x7b-v2-0-prometheus-8x7b-v2-0',
-    version: '1.4.0',
-    description:
-      'Prometheus 2 is an alternative of GPT-4 evaluation when doing',
-    tags: [
-      'LAB starter',
-      'Task_name 1',
-      'Task_name 2',
-      'Task_name 3',
-      'Task_name 4',
-    ],
-  },
-  {
-    title: 'granite-7b-starter',
-    version: '1.4.0',
-    description: 'Custom Red Hat phase00 tuned base model.',
-    tags: [
-      'LAB starter',
-      'Task_name 1',
-      'Task_name 2',
-      'Task_name 3',
-      'Task_name 4',
-    ],
-  },
-  {
-    title: 'granite-7b-starter',
-    version: '1.4.0',
-    description: 'Custom Red Hat phase00 tuned base model.',
-    tags: [
-      'LAB starter',
-      'Task_name 1',
-      'Task_name 2',
-      'Task_name 3',
-      'Task_name 4',
-    ],
-  },
-];
-
-export const AI_MODELS_DESCRIPTION =
-  'Discover the AI models and services that are available in your organization';
-
-export const AI_TEMPLATES = [
-  {
-    title: 'Chatbot template',
-    description: `Pre-built chatbot with NLP capabilities using models like OpenAI's GPT or Google's Dialogflow for customer support Pre-built chatbot with NLP capabilities using models like OpenAI's GPT or Google's Dialogflow for customer support...`,
-    tag: 'Template',
-  },
-  {
-    title: 'AI_Image_Recognition_Template_AI_Image-Recognition_Template',
-    description: `Uses computer vision models like OpenCV, TensorFlow, or YOLO to detect and classify objects in images or videos. Uses computer vision models like OpenCV, TensorFlow, or YOLO to detect and classify objects in images or videos.`,
-    tag: 'Template',
-  },
-  {
-    title: 'AI Recommendation System',
-    description: `Includes collaborative filtering and deep learning models (e.g., TensorFlow Recommenders) to provide personalized content or...`,
-    tag: 'Service',
-  },
-  {
-    title: 'AI Speech-to-Text Template',
-    description: `Leverages ASR models like OpenAI Whisper or Google's Speech-to-Text API for transcribing audio into text in...`,
-    tag: 'Service',
-  },
-  {
-    title: 'Chatbot template',
-    description: `Pre-built chatbot with NLP capabilities using models like OpenAI's GPT or Google's Dialogflow for customer support Pre-built chatbot with NLP capabilities using models like OpenAI's GPT or Google's Dialogflow for customer support...`,
-    tag: 'Template',
-  },
-  {
-    title: 'Chatbot template',
-    description: `Pre-built chatbot with NLP capabilities using models like OpenAI's GPT or Google's Dialogflow for customer support...`,
-    tag: 'Template',
+    buttonLink: '/learning-paths',
   },
 ];